Finalized Feeder for import into master
