+++ src/main/java/frc/robot/Robot.java
@@ -7,33 +7,19 @@
 
 package frc.robot;
 
-import com.revrobotics.CANSparkMax;
-import com.revrobotics.CANSparkMaxLowLevel.MotorType;
-
 import edu.wpi.first.wpilibj.TimedRobot;
-import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.RobotContainer;
-import frc.robot.commands.C_FeederDefault;
-import frc.robot.test.C_FeederTest;
-import frc.robot.test.C_TestShoot;
-import frc.robot.drivers.TimeOfFlightSensor;
-import frc.robot.subsystems.SS_Feeder;
 
 public class Robot extends TimedRobot {
   private final RobotContainer container = new RobotContainer();
   
-  // private CANSparkMax beltMotor = new CANSparkMax(Constants.FEED_MOTOR_CANID, MotorType.kBrushless);
-  // private TimeOfFlightSensor entrySensor = new TimeOfFlightSensor(Constants.ENTRY_SENSOR_CANID);
-  // private TimeOfFlightSensor exitSensor = new TimeOfFlightSensor(Constants.EXIT_SENSOR_CANID);
-  // private SS_Feeder feeder = new SS_Feeder(beltMotor, entrySensor, exitSensor);
   @Override
   public void robotInit() {
   }
 
   @Override
   public void robotPeriodic() {
-    //CommandScheduler.getInstance().schedule(new C_TestShoot(shooter, feeder));
     CommandScheduler.getInstance().run();
   }
 }
