+++ src/main/java/frc/robot/commands/C_PrepIntakeShoot.java
@@ -5,42 +5,37 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.test;
+package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.SS_Feeder;
-import frc.robot.subsystems.SS_Shooter;
 import frc.robot.subsystems.SS_Feeder.FeedMode;
 
-public class C_TestShoot extends CommandBase {
+public class C_PrepIntakeShoot extends CommandBase {
   /**
-   * Creates a new C_TestShoot.
+   * Creates a new C_PrepIntakeShoot.
    */
-  private SS_Shooter shooter;
   private SS_Feeder feeder;
-  public C_TestShoot(SS_Shooter shooter, SS_Feeder feeder) {
-    // Use addRequirements() here to declare subsystem dependencies.
+  public C_PrepIntakeShoot(SS_Feeder feeder) {
     this.feeder = feeder;
-    this.shooter = shooter;
-    addRequirements(shooter, feeder);
+    addRequirements(feeder);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    shooter.testSetTargetRPM(200);
-    shooter.setSpinning(true);
+   feeder.setFeedMode(FeedMode.PRESHOOT);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    feeder.setFeedMode(FeedMode.SHOOT);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
+    feeder.setFeedMode(FeedMode.STOPPED);
   }
 
   // Returns true when the command should end.
