--- src/main/java/frc/robot/test/C_PrepShoot.java
@@ -1,50 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.test;
-
-import org.frcteam2910.common.robot.input.Controller;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.SS_Shooter;
-
-public class C_PrepShoot extends CommandBase {
-  /**
-   * Creates a new C_PrepShoot.
-   */
-  private SS_Shooter shooter;
-  public C_PrepShoot(SS_Shooter shooter) {
-    this.shooter = shooter;
-    addRequirements(shooter);
-    // Use addRequirements() here to declare subsystem dependencies.
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    shooter.testSetTargetRPM(1000);
-    shooter.setSpinning(true);
-    //shooter.updateFromVision(true).setSpinning(true);
-    //shooter.setHoodFar(true);
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return true;
-  }
-}
